sort circular dependencies :(
