+++ src/twisted/python/util.py
@@ -14,11 +14,10 @@ try:
 except ImportError:
     setgroups = getgroups = None
 
-from functools import wraps
-
 from twisted.python.compat import _PY3, unicode
 from incremental import Version
 from twisted.python.deprecate import deprecatedModuleAttribute
+from twisted.python._oldstyle import _oldStyle, _replaceIf
 
 # For backwards compatibility, some things import this, so just link it
 from collections import OrderedDict
@@ -996,37 +995,6 @@ def runWithWarningsSuppressed(suppressedWarnings, f, *args, **kwargs):
 
 
 
-def _replaceIf(condition, alternative):
-    """
-    If C{condition}, replace this function with C{alternative}.
-
-    @param condition: A L{bool} which says whether this should be replaced.
-
-    @param alternative: An alternative function that will be swapped in instead
-        of the original, if C{condition} is truthy.
-
-    @return: A decorator.
-    """
-    def decorator(func):
-
-        if condition is True:
-            call = alternative
-        elif condition is False:
-            call = func
-        else:
-            raise ValueError(("condition argument to _replaceIf requires a "
-                              "bool, not {}").format(repr(condition)))
-
-        @wraps(func)
-        def wrapped(*args, **kwargs):
-            return call(*args, **kwargs)
-
-        return wrapped
-
-    return decorator
-
-
-
 __all__ = [
     "uniquify", "padTo", "getPluginDirs", "addPluginDir", "sibpath",
     "getPassword", "println", "makeStatBar", "OrderedDict",
@@ -1034,7 +1002,7 @@ __all__ = [
     "raises", "IntervalDifferential", "FancyStrMixin", "FancyEqMixin",
     "switchUID", "SubclassableCStringIO", "mergeFunctionMetadata",
     "nameToLabel", "uidFromString", "gidFromString", "runAsEffectiveUser",
-    "untilConcludes", "runWithWarningsSuppressed",
+    "untilConcludes", "runWithWarningsSuppressed", "_replaceIf",
 ]
